requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.42
psycopg2-binary==2.9.10
pydantic-settings==2.10.1
python-dotenv==1.1.1
alembic==1.16.4